CustomerCreator Revamp
Changed CustomerCreator so it is compatible w/ the clock class and fits with Dr.Wolff's description of the class.  Her definition of this class: " An object of this class is passed the minimum and maximum interarrival and service times of the customers upon instantiation. Its primary responsibility is to generate and return the “next” customer when requested."