New time mechanic and output
I've started working on implementing time into our simulator class. Note: the time is not using the clock method just yet; it will in the future. I've started working on the output and added a method in LinkedListQueue that gets the lastly added node/customer object in the list. Finally, I've added a "pre-queue" arraylist where new customer objects are stored. This will come in handy for the time mechanic.